+++ extensions/vscode/config_schema.json
@@ -2133,6 +2133,29 @@
             "promptPath": {
               "type": "string"
             },
+            "quickActions": {
+              "type": "array",
+              "items": {
+                "type": "object",
+                "properties": {
+                  "title": {
+                    "type": "string",
+                    "description": "The title of the quick action that will display in the Code Lens."
+                  },
+                  "prompt": {
+                    "type": "string",
+                    "description": "The prompt that will be sent to the model when the quick action is invoked, with the function or class body concatenated."
+                  },
+                  "sendToChat": {
+                    "type": "boolean",
+                    "description": "If true, the result of the quick action will be sent to the chat panel. If false, the streamed result will be inserted into the document.",
+                    "default": false
+                  }
+                },
+                "required": ["prompt", "title"]
+              },
+              "description": "Quick actions are a way to add custom commands to the Code Lens of function and class declarations."
+            },
             "contextMenuPrompts": {
               "type": "object",
               "properties": {
