+++ extensions/vscode/src/diff/verticalPerLine/manager.ts
@@ -172,11 +172,36 @@ export class VerticalPerLineDiffManager {
     }
   }
 
+  /**
+   * Streams an edit to the current document based on user input and model output.
+   *
+   * @param input - The user's input or instruction for the edit.
+   * @param modelTitle - The title of the language model to be used.
+   * @param [onlyOneInsertion] - Optional flag to limit the edit to a single insertion.
+   * @param [quickEdit] - Optional string indicating if this is a quick edit.
+   * @param [range] - Optional range to use instead of the highlighted text. Note that the `quickEdit`
+   *                  property currently can't be passed with `range` since it assumes there is an
+   *                  active selection.
+   *
+   * This method performs the following steps:
+   * 1. Sets up the editor context for the diff.
+   * 2. Determines the range of text to be edited.
+   * 3. Clears any existing diff handlers for the file.
+   * 4. Creates a new vertical diff handler.
+   * 5. Prepares the context (prefix and suffix) for the language model.
+   * 6. Streams the diff lines from the language model.
+   * 7. Applies the changes to the document.
+   * 8. Sets up a listener for subsequent user edits.
+   *
+   * The method handles various edge cases, such as quick edits and existing diffs,
+   * and manages the lifecycle of diff handlers and document change listeners.
+   */
   async streamEdit(
     input: string,
     modelTitle: string | undefined,
     onlyOneInsertion?: boolean,
     quickEdit?: string,
+    range?: vscode.Range,
   ) {
     vscode.commands.executeCommand("setContext", "continue.diffVisible", true);
 
@@ -188,9 +213,15 @@ export class VerticalPerLineDiffManager {
 
     const filepath = editor.document.uri.fsPath;
 
-    // Initialize start/end at editor selection
-    let startLine = editor.selection.start.line;
-    let endLine = editor.selection.end.line;
+    let startLine, endLine: number;
+
+    if (range) {
+      startLine = range.start.line;
+      endLine = range.end.line;
+    } else {
+      startLine = editor.selection.start.line;
+      endLine = editor.selection.end.line;
+    }
 
     // Check for existing handlers in the same file the new one will be created in
     const existingHandler = this.getHandlerForFile(filepath);
@@ -272,11 +303,13 @@ export class VerticalPerLineDiffManager {
       llm.model,
     );
 
-    // Unselect the range
-    editor.selection = new vscode.Selection(
-      editor.selection.active,
-      editor.selection.active,
-    );
+    if (editor.selection) {
+      // Unselect the range
+      editor.selection = new vscode.Selection(
+        editor.selection.active,
+        editor.selection.active,
+      );
+    }
 
     vscode.commands.executeCommand(
       "setContext",
