--- extensions/vscode/src/lang-server/codeLens.ts
@@ -1,481 +1,0 @@
-import * as os from "node:os";
-import * as path from "node:path";
-import * as vscode from "vscode";
-import { DIFF_DIRECTORY, type DiffManager } from "../diff/horizontal";
-import type { VerticalDiffCodeLens } from "../diff/verticalPerLine/manager";
-import { editorSuggestionsLocked, editorToSuggestions } from "../suggestions";
-import { getAltOrOption, getMetaKeyLabel, getPlatform } from "../util/util";
-import { getExtensionUri } from "../util/vscode";
-import { ContinueConfig } from "core";
-
-class VerticalPerLineCodeLensProvider implements vscode.CodeLensProvider {
-  private _eventEmitter: vscode.EventEmitter<void> =
-    new vscode.EventEmitter<void>();
-  onDidChangeCodeLenses: vscode.Event<void> = this._eventEmitter.event;
-
-  public refresh(): void {
-    this._eventEmitter.fire();
-  }
-
-  constructor(
-    private readonly editorToVerticalDiffCodeLens: Map<
-      string,
-      VerticalDiffCodeLens[]
-    >,
-  ) {}
-
-  public provideCodeLenses(
-    document: vscode.TextDocument,
-    _: vscode.CancellationToken,
-  ): vscode.CodeLens[] | Thenable<vscode.CodeLens[]> {
-    const filepath = document.uri.fsPath;
-    const blocks = this.editorToVerticalDiffCodeLens.get(filepath);
-    if (!blocks) {
-      return [];
-    }
-
-    const codeLenses: vscode.CodeLens[] = [];
-    for (let i = 0; i < blocks.length; i++) {
-      const block = blocks[i];
-      const start = new vscode.Position(block.start, 0);
-      const range = new vscode.Range(
-        start,
-        start.translate(block.numGreen + block.numRed),
-      );
-      if (codeLenses.length === 0) {
-        codeLenses.push(
-          new vscode.CodeLens(range, {
-            title: `Accept All (${getMetaKeyLabel()}⇧↩)`,
-            command: "continue.acceptDiff",
-            arguments: [filepath, i],
-          }),
-          new vscode.CodeLens(range, {
-            title: `Reject All (${getMetaKeyLabel()}⇧⌫)`,
-            command: "continue.rejectDiff",
-            arguments: [filepath, i],
-          }),
-        );
-      }
-      codeLenses.push(
-        new vscode.CodeLens(range, {
-          title: `Accept${
-            codeLenses.length === 2
-              ? ` (${getAltOrOption()}${getMetaKeyLabel()}Y)`
-              : ""
-          }`,
-          command: "continue.acceptVerticalDiffBlock",
-          arguments: [filepath, i],
-        }),
-        new vscode.CodeLens(range, {
-          title: `Reject${
-            codeLenses.length === 2
-              ? ` (${getAltOrOption()}${getMetaKeyLabel()}N)`
-              : ""
-          }`,
-          command: "continue.rejectVerticalDiffBlock",
-          arguments: [filepath, i],
-        }),
-      );
-      if (codeLenses.length === 4) {
-        codeLenses.push(
-          new vscode.CodeLens(range, {
-            title: `${getMetaKeyLabel()}I to add instructions`,
-            command: "",
-          }),
-        );
-      }
-    }
-
-    return codeLenses;
-  }
-}
-
-class SuggestionsCodeLensProvider implements vscode.CodeLensProvider {
-  public provideCodeLenses(
-    document: vscode.TextDocument,
-    _: vscode.CancellationToken,
-  ): vscode.CodeLens[] | Thenable<vscode.CodeLens[]> {
-    const suggestions = editorToSuggestions.get(document.uri.toString());
-    if (!suggestions) {
-      return [];
-    }
-    const locked = editorSuggestionsLocked.get(document.uri.fsPath.toString());
-
-    const codeLenses: vscode.CodeLens[] = [];
-    for (const suggestion of suggestions) {
-      const range = new vscode.Range(
-        suggestion.oldRange.start,
-        suggestion.newRange.end,
-      );
-      codeLenses.push(
-        new vscode.CodeLens(range, {
-          title: "Accept",
-          command: "continue.acceptSuggestion",
-          arguments: [suggestion],
-        }),
-        new vscode.CodeLens(range, {
-          title: "Reject",
-          command: "continue.rejectSuggestion",
-          arguments: [suggestion],
-        }),
-      );
-      if (codeLenses.length === 2) {
-        codeLenses.push(
-          new vscode.CodeLens(range, {
-            title: `(${getMetaKeyLabel()}⇧⏎/${getMetaKeyLabel()}⇧⌫ to accept/reject all)`,
-            command: "",
-          }),
-        );
-      }
-    }
-
-    return codeLenses;
-  }
-}
-
-class DiffViewerCodeLensProvider implements vscode.CodeLensProvider {
-  diffManager: DiffManager;
-
-  constructor(diffManager: DiffManager) {
-    this.diffManager = diffManager;
-  }
-
-  public provideCodeLenses(
-    document: vscode.TextDocument,
-    _: vscode.CancellationToken,
-  ): vscode.CodeLens[] | Thenable<vscode.CodeLens[]> {
-    if (path.dirname(document.uri.fsPath) === DIFF_DIRECTORY) {
-      const codeLenses: vscode.CodeLens[] = [];
-      let range = new vscode.Range(0, 0, 1, 0);
-      const diffInfo = this.diffManager.diffAtNewFilepath(document.uri.fsPath);
-      if (diffInfo) {
-        range = diffInfo.range;
-      }
-      codeLenses.push(
-        new vscode.CodeLens(range, {
-          title: `Accept All ✅ (${getMetaKeyLabel()}⇧⏎)`,
-          command: "continue.acceptDiff",
-          arguments: [document.uri.fsPath],
-        }),
-        new vscode.CodeLens(range, {
-          title: `Reject All ❌ (${getMetaKeyLabel()}⇧⌫)`,
-          command: "continue.rejectDiff",
-          arguments: [document.uri.fsPath],
-        }),
-        // new vscode.CodeLens(range, {
-        //   title: `Further Edit ✏️ (${getMetaKeyLabel()}⇧M)`,
-        //   command: "continue.focusContinueInputWithEdit",
-        // })
-      );
-      return codeLenses;
-    } else {
-      return [];
-    }
-  }
-}
-
-class ConfigPyCodeLensProvider implements vscode.CodeLensProvider {
-  public provideCodeLenses(
-    document: vscode.TextDocument,
-    _: vscode.CancellationToken,
-  ): vscode.CodeLens[] | Thenable<vscode.CodeLens[]> {
-    const codeLenses: vscode.CodeLens[] = [];
-
-    if (
-      !document.uri.fsPath.endsWith(".continue/config.json") &&
-      !document.uri.fsPath.endsWith(".continue\\config.json")
-    ) {
-      return codeLenses;
-    }
-
-    const lines = document.getText().split(os.EOL);
-    const lineOfModels = lines.findIndex((line) =>
-      line.includes('"models": ['),
-    );
-
-    if (lineOfModels >= 0) {
-      const range = new vscode.Range(lineOfModels, 0, lineOfModels + 1, 0);
-      // codeLenses.push(
-      //   new vscode.CodeLens(range, {
-      //     title: `+ Add a Model`,
-      //     command: "continue.addModel",
-      //   })
-      // );
-    }
-
-    const lineOfSystemMessage = lines.findIndex((line) =>
-      line.includes("ContinueConfig("),
-    );
-
-    if (lineOfSystemMessage >= 0) {
-      const range = new vscode.Range(
-        lineOfSystemMessage,
-        0,
-        lineOfSystemMessage + 1,
-        0,
-      );
-      codeLenses.push(
-        new vscode.CodeLens(range, {
-          title: "✏️ Edit in UI",
-          command: "continue.openSettingsUI",
-        }),
-      );
-    }
-
-    return codeLenses;
-  }
-}
-
-interface TutorialCodeLensItems {
-  lineIncludes: string;
-  commands: vscode.Command[];
-}
-
-const cmdCtrl = getPlatform() === "mac" ? "Cmd" : "Ctrl";
-
-const actions: TutorialCodeLensItems[] = [
-  {
-    lineIncludes: `Step 2: Use the keyboard shortcut [${cmdCtrl}+L]`,
-    commands: [
-      {
-        title: `${cmdCtrl}+L`,
-        command: "continue.focusContinueInput",
-      },
-    ],
-  },
-  {
-    lineIncludes: "Step 3: Ask a question",
-    commands: [
-      {
-        title: `"what does this code do?"`,
-        command: "continue.sendMainUserInput",
-        arguments: ["what does this code do?"],
-      },
-      {
-        title: `"what is an alternative to this?"`,
-        command: "continue.sendMainUserInput",
-        arguments: ["what is an alternative to this?"],
-      },
-    ],
-  },
-  {
-    lineIncludes: `Step 2: Use the keyboard shortcut [${cmdCtrl}+I] to edit`,
-    commands: [
-      {
-        title: `${cmdCtrl}+I`,
-        command: "continue.quickEdit",
-        arguments: ["Add comments"],
-      },
-    ],
-  },
-  {
-    lineIncludes: "Step 1: Run this Python file",
-    commands: [
-      {
-        title: "Run the file",
-        command: "continue.sendToTerminal",
-        arguments: [
-          `python ${path.join(
-            getExtensionUri().fsPath,
-            "continue_tutorial.py",
-          )}\n`,
-        ],
-      },
-    ],
-  },
-  {
-    lineIncludes: "Step 2: Use the keyboard shortcut cmd/ctrl + shift + R",
-    commands: [
-      {
-        title: "Debug the error",
-        command: "continue.debugTerminal",
-      },
-    ],
-  },
-  {
-    lineIncludes: `Step 2: Use the keyboard shortcut [${cmdCtrl}+Shift+R]`,
-    commands: [
-      {
-        title: `${cmdCtrl}+Shift+R`,
-        command: "continue.debugTerminal",
-      },
-    ],
-  },
-];
-
-class TutorialCodeLensProvider implements vscode.CodeLensProvider {
-  public provideCodeLenses(
-    document: vscode.TextDocument,
-    _: vscode.CancellationToken,
-  ): vscode.CodeLens[] | Thenable<vscode.CodeLens[]> {
-    const codeLenses: vscode.CodeLens[] = [];
-
-    if (!document.uri.fsPath.endsWith("continue_tutorial.py")) {
-      return codeLenses;
-    }
-
-    const lines = document.getText().split(os.EOL);
-
-    for (const action of actions) {
-      const lineOfAction = lines.findIndex((line) =>
-        line.includes(action.lineIncludes),
-      );
-
-      if (lineOfAction >= 0) {
-        const range = new vscode.Range(lineOfAction, 0, lineOfAction + 1, 0);
-        for (const command of action.commands) {
-          codeLenses.push(new vscode.CodeLens(range, command));
-        }
-      }
-    }
-
-    const lineOf11 = lines.findIndex((line) =>
-      line.includes("Step 1: Highlight the function below"),
-    );
-    if (lineOf11 >= 0) {
-      const range = new vscode.Range(lineOf11, 0, lineOf11 + 1, 0);
-      codeLenses.push(
-        new vscode.CodeLens(range, {
-          title: "Highlight the function",
-          command: "continue.selectRange",
-          arguments: [lineOf11 + 3, lineOf11 + 11],
-        }),
-      );
-    }
-    const lineOf21 = lines.findIndex((line) =>
-      line.includes("Step 1: Highlight this code"),
-    );
-    if (lineOf21 >= 0) {
-      const range = new vscode.Range(lineOf21, 0, lineOf21 + 1, 0);
-      codeLenses.push(
-        new vscode.CodeLens(range, {
-          title: "Highlight the function",
-          command: "continue.selectRange",
-          arguments: [lineOf21 + 3, lineOf21 + 14],
-        }),
-      );
-    }
-
-    return codeLenses;
-  }
-}
-
-class QuickActionsLensProvider implements vscode.CodeLensProvider {
-  static quickActionSymbolKinds = [
-    vscode.SymbolKind.Function,
-    vscode.SymbolKind.Class,
-  ];
-
-  async provideCodeLenses(
-    document: vscode.TextDocument,
-  ): Promise<vscode.CodeLens[]> {
-    const symbols = await vscode.commands.executeCommand<
-      Array<vscode.DocumentSymbol | vscode.SymbolInformation>
-    >("vscode.executeDocumentSymbolProvider", document.uri);
-
-    const filteredSmybols = symbols?.filter((def) =>
-      QuickActionsLensProvider.quickActionSymbolKinds.includes(def.kind),
-    );
-
-    const explain: vscode.Command = {
-      command: "continue.explain",
-      title: "Continue: Explain",
-    };
-
-    const comment: vscode.Command = {
-      command: "continue.writeDocstringForCode",
-      title: "Comment",
-    };
-
-    const disable: vscode.Command = {
-      command: "continue.disableQuickActions",
-      title: "X",
-    };
-
-    let codeLenses = filteredSmybols.flatMap((def: any) => [
-      new vscode.CodeLens(def.range, {
-        ...explain,
-        arguments: [document.getText(def.range)],
-      }),
-      new vscode.CodeLens(def.range, comment),
-      new vscode.CodeLens(def.range, disable),
-    ]);
-
-    return codeLenses;
-  }
-}
-
-export let verticalPerLineCodeLensProvider: vscode.Disposable | undefined =
-  undefined;
-let diffsCodeLensDisposable: vscode.Disposable | undefined = undefined;
-let suggestionsCodeLensDisposable: vscode.Disposable | undefined = undefined;
-let configPyCodeLensDisposable: vscode.Disposable | undefined = undefined;
-let tutorialCodeLensDisposable: vscode.Disposable | undefined = undefined;
-let quickActionsCodeLensDisposable: vscode.Disposable | undefined = undefined;
-
-export function registerAllCodeLensProviders(
-  context: vscode.ExtensionContext,
-  diffManager: DiffManager,
-  editorToVerticalDiffCodeLens: Map<string, VerticalDiffCodeLens[]>,
-  config: ContinueConfig,
-) {
-  if (verticalPerLineCodeLensProvider) {
-    verticalPerLineCodeLensProvider.dispose();
-  }
-  if (suggestionsCodeLensDisposable) {
-    suggestionsCodeLensDisposable.dispose();
-  }
-  if (diffsCodeLensDisposable) {
-    diffsCodeLensDisposable.dispose();
-  }
-  if (configPyCodeLensDisposable) {
-    configPyCodeLensDisposable.dispose();
-  }
-  if (tutorialCodeLensDisposable) {
-    tutorialCodeLensDisposable.dispose();
-  }
-  if (quickActionsCodeLensDisposable) {
-    quickActionsCodeLensDisposable.dispose();
-  }
-
-  const verticalDiffCodeLens = new VerticalPerLineCodeLensProvider(
-    editorToVerticalDiffCodeLens,
-  );
-  verticalPerLineCodeLensProvider = vscode.languages.registerCodeLensProvider(
-    "*",
-    verticalDiffCodeLens,
-  );
-  suggestionsCodeLensDisposable = vscode.languages.registerCodeLensProvider(
-    "*",
-    new SuggestionsCodeLensProvider(),
-  );
-  diffsCodeLensDisposable = vscode.languages.registerCodeLensProvider(
-    "*",
-    new DiffViewerCodeLensProvider(diffManager),
-  );
-  configPyCodeLensDisposable = vscode.languages.registerCodeLensProvider(
-    "*",
-    new ConfigPyCodeLensProvider(),
-  );
-  tutorialCodeLensDisposable = vscode.languages.registerCodeLensProvider(
-    "*",
-    new TutorialCodeLensProvider(),
-  );
-
-  if (!!config.experimental?.quickActions) {
-    quickActionsCodeLensDisposable = vscode.languages.registerCodeLensProvider(
-      "*",
-      new QuickActionsLensProvider(),
-    );
-
-    context.subscriptions.push(quickActionsCodeLensDisposable);
-  }
-
-  context.subscriptions.push(verticalPerLineCodeLensProvider);
-  context.subscriptions.push(suggestionsCodeLensDisposable);
-  context.subscriptions.push(diffsCodeLensDisposable);
-  context.subscriptions.push(configPyCodeLensDisposable);
-  context.subscriptions.push(tutorialCodeLensDisposable);
-
-  return { verticalDiffCodeLens, quickActionsCodeLensDisposable };
-}
