+++ extensions/vscode/src/lang-server/codeLens/index.ts
@@ -1,0 +1,1 @@
+export { registerAllCodeLensProviders } from "./registerAllCodeLensProviders";
