+++ extensions/vscode/src/lang-server/codeLens/providers/index.ts
@@ -1,0 +1,6 @@
+export { ConfigPyCodeLensProvider } from "./ConfigPyCodeLensProvider";
+export { DiffViewerCodeLensProvider } from "./DiffViewerCodeLensProvider";
+export { QuickActionsCodeLensProvider } from "./QuickActionsCodeLensProvider";
+export { SuggestionsCodeLensProvider } from "./SuggestionsCodeLensProvider";
+export { TutorialCodeLensProvider } from "./TutorialCodeLensProvider";
+export { VerticalPerLineCodeLensProvider } from "./VerticalPerLineCodeLensProvider";
